Update Login Update Logout Delete

+++ src/main/java/com/mediquick/web/primary/role/domain/RoleRepository.java
@@ -2,6 +2,6 @@ package com.mediquick.web.primary.role.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+public interface RoleRepository extends JpaRepository<Role, Byte> {
     public Role findByCode(Byte code);
 }
